Add UI to messaging sample (#52)
This PR adds a simple UITableView UI to the Messaging_Example app, which shows the status of the APNs token and the FCM token. Both the APNs token and the FCM token are sharable by tapping their respective row. This makes it easy to transfer the token to your computer to do some testing.

By default the app always registers for silent notifications (which don't require user permission), with a separate button to request user-visible notifications (requiring user permission). It reflects the types of notifications (silent, alert, badge, sounds) that have been granted by the user.

It also informs the user that APNs does not work in a Simulator build, as that is not supported by Apple.