Add GoogleService-Info.plist check to Messaging sample (#59)
This adds a new class, FIRSampleAppUtilities, which is used by the Messaging sample app to skip calling FirebaseApp.configure() if the plist file is missing or is the dummy file in the repo. It encourages the developer to replace it with a valid plist file, before the sample app can be used.

The other sample apps haven't been updated to use this check, as they are not actually proper sample apps yet. When they actually get fleshed out, they should use the plist check in their respective AppDelegate classes.

NOTE: The alert to the developer tells them to visit https://github.com/firebase/firebase-ios-sdk/ which is not public as of today, but will be for I/O 2017.