Merge public 5 25 (#79)
* Initial commit

* Initial

* Add three files in repo despite .gitignore

* Update README with context, contributing, and license sections (#3)

* Updating READMEs (#5)

* Add comments to each FIRLoggerLevel value. (#6)

Docs aren't being generated properly for FIRLoggerLevel because
it's missing the proper documentation.

* Update travis for renamed AllUnitTests scheme (#8)

* Remove double commented lines. (#11)

* Update scheme setting in test.sh (#12)

* Update travis xcode version to 8.3 (#13)

* Add CI badge to README (#14)

* Fix typos (#15)

* Adding cache to Travis builds (#17)

* Enable direct channel in the sample app (#7)

This shows the simple property, shouldEstablishDirectChannel, to open a new channel, and how to show incoming messages with the iOS 10 delegate handler

* Auth jazzy fixes and Swift sample updates (#19)

* Removes the '@c' that would cause Jazzy issue down the road.

* Update travis test to retry if error is 65 (#20)

* Fixes ApiTests and separates its credentials from Sample (#25)

* Log an error if for some reason we can't add a method to a class (#24)

Maybe it was already swizzled, or some other issue. Either way, it would be good to log this error, as it might help us / developers diagnose an issue.

* Update FIRPhoneAuthProvider reference docs (#27)

Adds reference docs error documentation to FIRPhoneAuthProvider.
Also fixes small typo in FIRAuthErrorUtils.

* Clean up logging and configuration in Messaging (#28)

* Remove mostly unused code from FIRMessagingLogger

This cleans up some left-over old logging logic from before we moved to FIRLogger in Firebase Core. We no longer need this logging functionality.

* Delete FIRMessagingConfig.{h,m}

We no longer need a config class to store an unused log level filter (that filter is no longer needed).