Add section explaining community supported efforts
As we are adding in support for things like macOS which can't (yet) be officially supported, added a section in the README making clear that we want to allow community-support efforts like this to exist, but that these efforts don't imply or infer that Firebase will be officially providing documentation, technical support, etc. for these efforts.

Also demote the specific SDK sections by one level, making the README a bit more parseable.