Remove FIRMessaging_FAIL macro, log result code
This removes the `FIRMessaging_FAIL` macro which was using `__builtin_trap()`, and replaced with `NSAssert` calls. These `NSAssert` calls may not get called in release builds, and so we also log them with FIRLogger error messages.

The RMQ database open error result code is now parsed and included in the error message to help us identify causes for #199