Remove FIRAppEnvironmentUtil's sharedApplication (#603)
* Deprecate FIRAppEnvironmentUtil sharedApplication

* Remove sharedApplication private API.

This also allows us to remove an AppKit dependency.

* Disable autolink when building FirebaseCore (#606)

This prevents the dependencies of FirebaseCore's dependencies from
bleeding into its interface.

The proximate benefit of avoiding this is that FirebaseCore no longer
ends up depending on ColorSync.framework which became a top-level system
framework in 10.13. This makes it possible to build against the
resulting FirebaseCore.framework on macOS 10.12 using Xcode 9 (with the
macOS 10.13 SDK).