GetOptions for controlling offline behaviour.
Add option to allow the user to control where
DocumentReference.getDocument() fetches from. By default, it fetches
from the server (if possible) and falls back to the local cache. It's
now possible to alternatively fetch from the local cache only, or to
fetch from the server only (though in the server only case, latency
compensation is still enabled).

Still TODO: Also enable this for CollectionReferenece.getDocuments().