Fix test failures that occur during prod build (#910)
* Fix ArraySortedMap.ChecksSize test (when NDEBUG is enabled)

assert() is a noop when NDEBUG is defined (i.e. during production
builds) so the size assertion check doesn't occur. Assuming we want this
check even during prod use, I've switched assert to FIREBASE_ASSERT.

* s/FIREBASE_ASSERT/FIREBASE_DEV_ASSERT/g in ordered_code.cc

This keeps the behaviour in line with the original (which purposely
differs if NDEBUG is defined or not.) Discovered due to the test suite
(correctly) breaking during a prod build.