Firestore project: Import of API interfaces.
Manually import the public portion of

  - 22c226af3f5570514d3d13d82a399577ecd7d280 C++ migration: make Timestamp class a part of public API ... by Konstantin Varlamov <var-const@users.noreply.github.com>

Note that this only imports the header, which otherewise won't come over with a
regular copybara run because we're currently excluding the public headers from
the Firestore import with cl/188810622.

The .cc implementation of this will come over with a regular copybara import,
coming shortly.

ORIGINAL_AUTHOR=Gil <wilhuff@github.com>
PiperOrigin-RevId: 190986604