Fix iteration past end in LevelDBTransaction::Iterator (#1105)
Change Next() to avoid advancing past the end of either iterator by only
advancing if !advanced and if the iterator is still valid.

If Next() is called after the last entry is deleted, SyncToTransaction()
returns false indicating it hasn't advanced but at that point is_valid_
is false. Previously Next() would then attempt to advance one of the
underlying iterators and if it was the mutations_iter_ would advance
past the end, which would cause the next UpdateCurrent() to dereference
random memory.