Add AsyncQueue class.
AsyncQueue is a queue that executes given operations asynchronously,
enforcing that only a single operation is executing at any given time,
and that in-progress operations don't spawn more operations. The actual
execution is delegated to a platform-specific executor.

Also add "prototypes" of tests for AsyncQueue, to be instantiated with
concrete Executors in follow-up commits.