Backport array contains / transform improvements from Web.
This brings iOS into parity with the result of https://github.com/firebase/firebase-js-sdk/pull/763.

NOTE: I ported awaitLocalEventWithName / awaitRemoteEventWithName to FSTEventAccumulator but actually getting rid of the helpers in FIRArrayTransformTests and FIRServerTransformTests would have added significant boilerplate (because FSTEventAccumulator makes you pass a name for the events it awaits). So it's only a half-hearted port and I'm half tempted to revert it.