Temporarily disable Messaging pod lints
There's an unreleased change in Core that Messaging relies on, but Travis currently pulls in the publicly released version of Core causing the linting to fail. This should be re-enabled once either a) Core is released, or b) we support pointing to local pods for dependencies on Travis - whichever of the two come first.

An issue will be opened to re-enable the linting.