local::SerializerTest -> local::LocalSerializerTest
It turns out xcode also gets confused if two test fixtures have the same
name (even if they're in independent tests) and will pick one of them to
use for both sets of tests.  :(

Renamed test fixture to get around this. (It's more fitting with the
rename of the class under test, so even makes sense independently of
xcode.)