Cherry pick fix for #1591 for M30 (#1626)
* Fixes decoding WriteResults for deletes (with no updateTime). (#1622)

This should resolve #1591.  We were inadvertantly decoding the WriteResult as
having an updateTime of 0 which caused it to be older than the version of the
document in the RemoteDocumentsCache. Therefore we kept the old version in
the cache instead of persisting the deleted document. This caused the document
to reappear in the query results until we resolved it via a limbo resolution.

* Add changelog entry for delete WriteResult fix (d9f116). (#1623)

* Actually this fix is going out with 0.13.0