Add Status::CausedBy (#1635)
This makes it easy to wrap an underlying error in a higher-level one,
similar to the way Java exception chaining works.

Note that this intentionally diverges from upstream. The intent is that
this expresses a high-level concern so we can migrate to some common
status if one is ever released.