Fix crash in `ExecutorLibdispatch` when running unit tests (#1772)
Make the destructor of ExecutorLibdispatch turn any scheduled operations into no-ops (because whatever is scheduled in libdispatch cannot be truly canceled). Under normal conditions, Firestore should never trigger a situation when an executor is being destroyed while it still has pending operations. However, this can be triggered when running unit tests by just setting a breakpoint -- it will lead to a TimeSlot's execution that tries to access the executor after it's already destroyed.