gRPC C++: don't wait for initial server metadata during `GrpcStream::Start` (#1786)
By default, StartCall in gRPC C++ client would wait until the initial server metadata is received, after which the corresponding tag will come back from the completion queue. It can alternatively be configured so that initial server metadata is only received together with the first write operation, effectively making StartCall a synchronous operation. The benefit is that Firestore streams won't have to wait for a network roundtrip before they consider themselves open. It is also consistent with the other platforms.