Avoid creating a std::map of incomplete type.
This results in undefined behaviour.
(http://www.drdobbs.com/the-standard-librarian-containers-of-inc/184403814)

This patch is mostly inconsequential, since we intend to replace
std::map with immutable::SortedMap, but it turns out SortedMap makes the
problem much more obvious and immediately fails the compilation. This
patch does the move now, to make the subsequent patches more obvious.