Make api::DocumentReference accept ParsedSetData / ParsedUserData. (#2854)
I think the intention was that the user_data.h types should be used as the representation of parsed user data that can be passed down from the public API to the api:: layer, so this PR "fixes" api::DocumentReference to accept ParsedSetData and ParsedUserData in SetData() / UpdateData() instead of taking the list of mutations. This is consistent with the approach I took in the api::WriteBatch port.

Side note: Prior to this change, SetData() / UpdateData() were actually identical methods just with different names.