In case user is nil, we don't generate FIRAuthResult.
I investigated all FIRAuthDataResult usages in the Frebase Auth. There are two distinct ways to generate it.

Method 1:
FIRAuthResultCallback completion = ^(FIRUser * _Nullable user,
                                     NSError * _Nullable error) {
  FIRAuthDataResult *result =
    [[FIRAuthDataResult alloc] initWithUser:user
			 additionalUserInfo:info];
  callback(result, error); // Some calls callback(result, nil);
}

Method 2:
FIRAuthResultCallback completion = ^(FIRUser * _Nullable user,
                                     NSError * _Nullable error) {
  FIRAuthDataResult *result = user ?
    [[FIRAuthDataResult alloc] initWithUser:user
			 additionalUserInfo:info] : nil;
  callback(result, error);
}

I believe the second way is better for two reasons:

1. We don't need to change the public API to make user a nullable object.

2. When user is null, this means an error is occurred. So in this case, calling callback(nil, error) is more correct than calling  callback(result, nil/error) where result has an empty user.

So in this commit, we change a couple of places which uses Method 2 to follow Moethod 1, which most of the other places are using.