Use more robust method for generating modulemaps for zip build.
Currently we rely on an `xcconfig` generated from CocoaPods that changed
between 1.5.3 and 1.6.0, causing our modulemap generation not to work on
1.6+. Instead of relying on this, we should use the more robust script
to read the podspec and include all libraries and frameworks specified.
Unfortunately there's no simple way to find out which repo each
installed pod came from, so we have to list all repos, parse the output,
then search the filesystem in order of the specified repos.