Fix Firestore post-install when run in bundler
Previously when you run bundle exec pod install, this was running
sync_project.rb, but it was using the system ruby to do so based on the
shebang line in the script.

This change ensures that even if the system ruby does not have cocoapods
and its dependencies installed, when running under bundler the script
still works.