release-6.6.0 fix #2 for GDTUploadCoordinator crashes (#3557)
* Make ivar references safer (#3555)

Fixes 3547, in theory. This code all goes away/changes with the sqlite refactor, so making it safe is probably sufficient for now. No one was able to reproduce locally, most likely because this is a lifecycle issue.

* Increment GDT to v1.1.1

GDTCCT has a dep on ~> 1.1, so this incremented version should get picked up.

* Actually use the strongified variables I made