Increment GDT's version and CHANGELOG (#3691)
* Increment GDT's version and CHANGELOG

* Wrap decoding targetToInflightPackages in a try/catch (#3689)

It's not apparent how a key could be stored with a nil object in the dictionary. Maybe it's possible that somehow a GDTUploadPackage fails to decode, but there aren't any indications that this is the case. This PR adds a check for contents of the dictionary, though encoding an empty dictionary should be fine. A try/catch is utilized to catch exceptions thrown in iOS's implementation of NSSecureCoding in NSMutableDictionary.