Make `Serializer` fully support converting `FieldValue`s (#3659)
This involved support for reference values and taking server timestamps into account (by crashing, because they are not supposed to ever be serialized).

The biggest change is from making every function that decodes a `FieldValue` a member function (as well as any function called by those functions). That is because decoding a reference requires comparing database IDs, and the database ID of this Firestore instance is stored on the `Serializer`.