Fixes for issues discovered in the Unity desktop build (#4795) (#4825)
Fixes for issues discovered in the Unity desktop build (#4795)

Ensure that Firestore properly shuts down when the Unity AppDomain is unloaded. Improper shutdown manifested as a crash the next time Firestore was used because the LevelDB database files were still locked by the prior instance.

AppDomain unload differs from typical iOS termination because we want Dispose to be synchronous so that we signal shutdown is complete by the time Dispose returns. Without this, there's a race between shutdown in the current domain and startup in the next domain. This poses a problem though because disposers are run on the main thread, and callbacks from Firestore's async operations also typically run on the main thread. This means that by default there's no great way to implement Dispose without deadlock.

This change makes it such that the `api::Firestore` destructor now synchronously terminates the instance. There are a few moving parts to making this work:

  * The AsyncQueue now extends `std::enable_shared_from_this` captures a `shared_ptr` to itself whenever it enqueues an asynchronous operation. This ensures that if an operation is enqueued and the user doesn't wait for it to complete, the queue won't deallocate itself before the operation completes.

  * Makes ExecutorStd handle the case where its destructor might be running on a thread that it owns. This can now happen if an AsyncQueue Operation is the last thing holding a reference to itself.

  * Adds a new, synchronous FirestoreClient::Terminate method that does not have a callback. This is used to avoid the deadlock inherent with cross-queue notifications when the caller may be trying to block synchronously in the queue to which the notification would go.

  * Changes scheduled operations to hold their their references to FirestoreClient via `weak_ptr`, to avoid accidentally preserving the life of the FirestoreClient past Terminate.

  * Makes the api::Firestore destructor call a new synchronous Terminate method.