Merge release 6.21.0 (#5178)
* Update versions for Release 6.21.0

* Libraries should save state if a completion block is requested (#5147) (#5148)

* Libraries should save state if a completion block is requested

Currently, the onComplete block of sendXEvent:onComplete: is invoked when the event bytes are written to disk. However, if the app crashes before a lifecycle event occurs (trigger GDTCORStorage to serialize to disk), that event if effectively lost because the metadata has all been lost. This change triggers the saving of state of GDTCORStorage and prioritizers (if they implement it) if an onComplete block has been given.

* Remove RC from M67 (#5157)

* Revert "Migrate from using IID SDK for Remote Config to the new FIS SDK (#5096)" (#5156) (#5159)

This reverts commit 887ff82494df27e33a1d2ab09e3c8f6a0df86863.

* Only delete events when package was successfully delivered (#5172) (#5173)

* Only delete events when package was successfully delivered

* Add unit test to check that retrying a package delivery doesn't delete an event file

Co-authored-by: Michael Haney <mikehaney24@users.noreply.github.com>
Co-authored-by: Maksym Malyhin <mmaksym@google.com>